requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
anyio==3.6.2
argilla==1.6.0
async-timeout==4.0.2
asyncssh==2.13.1
attrs==23.1.0
awscli==1.27.123
backoff==2.2.1
bcrypt==4.0.1
blessed==1.20.0
blinker==1.6.2
boto3==1.26.123
botocore==1.29.123
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.0.4
clickhouse-connect==0.5.22
colorama==0.4.4
colorful==0.5.5
commonmark==0.9.1
cryptography==40.0.2
dataclasses-json==0.5.7
Deprecated==1.2.13
distlib==0.3.6
docutils==0.16
duckdb==0.7.1
et-xmlfile==1.1.0
fastapi==0.95.1
filelock==3.12.0
Flask==2.3.1
frozenlist==1.3.3
fsspec==2023.4.0
google-api-core==2.11.0
google-auth==2.17.3
googleapis-common-protos==1.59.0
gptcache==0.1.21
gpustat==1.1
greenlet==2.0.2
grpcio==1.43.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.142
llama-cpp-python==0.1.39
llama-index==0.5.27
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
nvidia-ml-py==11.525.112
oauth2client==4.1.3
olefile==0.46
openai==0.27.5
openapi-schema-pydantic==1.2.4
opencensus==0.11.2
opencensus-context==0.1.3
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
paramiko==3.1.0
pdfminer.six==20221105
pendulum==2.1.2
Pillow==9.5.0
platformdirs==3.5.0
posthog==3.0.1
prettytable==3.7.0
prometheus-client==0.16.0
protobuf==3.20.3
psutil==5.9.5
PuLP==2.7.0
py-spy==0.3.14
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.10.7
Pygments==2.15.1
PyNaCl==1.5.0
pypandoc==1.11
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytzdata==2020.1
PyYAML==5.4.1
ray==2.3.0
redis==4.5.4
regex==2023.3.23
requests==2.29.0
rfc3986==1.5.0
rich==13.0.1
rsa==4.7.2
runhouse==0.0.5
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
six==1.16.0
skypilot==0.2.5
smart-open==6.3.0
sniffio==1.3.0
SQLAlchemy==1.4.47
sshfs==2023.4.1
sshtunnel==0.4.0
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.6.2
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
Werkzeug==2.3.2
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
